requirements.txt
#Core AI Framework - Updated for compatibility
langchain==0.2.1
langchain-core==0.2.3
langchain-community==0.2.1
langchain-openai==0.1.7
langgraph==0.1.1
httpx<0.28

#UI Framework
streamlit==1.38.0
streamlit-chat==0.1.1

#Data Processing & Management
pandas==2.2.2 # Updated to a more recent stable version for Python 3.11
numpy==1.26.4
openpyxl==3.1.2
faker==24.0.0 # Version compatible with pandas update

#API Integrations
requests==2.31.0
twilio==9.2.4
python-dotenv==1.0.1

#AI/ML
openai==1.30.1
tiktoken==0.7.0

#Utilities
arrow==1.3.0
schedule==1.2.2
pydantic<3,>=1  # More flexible pydantic version for broad compatibility

#Development & Testing
pytest==8.3.3
black==24.10.0
flake8==7.1.1

#Optional Enhancements
plotly==5.24.1
altair==5.4.1

#Fuzzy String Matching for Patient Lookup
thefuzz==0.22.1
python-Levenshtein==0.25.1